+++ lib/esp_provisioning_wifi.dart
@@ -4,5 +4,6 @@ export 'esp_provisioning_error_codes.dart';
 export 'esp_provisioning_event.dart';
 export 'esp_provisioning_service.dart';
 export 'esp_provisioning_state.dart';
+export 'esp_security_scheme.dart';
 export 'esp_wifi_network.dart';
 export 'src/flutter_esp_ble_prov/flutter_esp_ble_prov.dart';